refactor: move claimUserTicket file to its own file
